+++ utils/kafka-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/exception/KafkaErrorTopicInitializer.scala
@@ -1,18 +1,15 @@
-package pl.touk.nussknacker.engine.lite.kafka
+package pl.touk.nussknacker.engine.kafka.exception
 
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.kafka.clients.admin.NewTopic
-import pl.touk.nussknacker.engine.kafka.exception.KafkaExceptionConsumerConfig
 import pl.touk.nussknacker.engine.kafka.{KafkaConfig, KafkaUtils}
 
-import java.util
 import java.util.Optional
 import java.util.concurrent.TimeUnit
+import java.{lang, util}
 import scala.jdk.CollectionConverters.asScalaSetConverter
 
-class ErrorHandlerInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfig: KafkaExceptionConsumerConfig) extends LazyLogging {
-
-  private val partitions = 2
+class KafkaErrorTopicInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfig: KafkaExceptionConsumerConfig) extends LazyLogging {
 
   private val timeoutSeconds = 5
 
@@ -25,9 +22,9 @@ class ErrorHandlerInitializer(kafkaConfig: KafkaConfig, exceptionHandlerConfig: 
         case (true, _) =>
           logger.debug("Topic exists, skipping")
         case (false, true) =>
-          logger.info(s"Creating error topic: $errorTopic with default config: partitions: $partitions, check if the values are correct")
-          admin.createTopics(util.Arrays.asList(new NewTopic(errorTopic, Optional.of[java.lang.Integer](partitions), Optional.empty[java.lang.Short]())))
-            .all().get(timeoutSeconds, TimeUnit.SECONDS)
+          logger.info(s"Creating error topic: $errorTopic with default configs, please check if the values are correct")
+          val errorTopicConfig = new NewTopic(errorTopic, Optional.empty[Integer](), Optional.empty[lang.Short]())
+          admin.createTopics(util.Arrays.asList(errorTopicConfig)).all().get(timeoutSeconds, TimeUnit.SECONDS)
         case (false, false) =>
           logger.error(s"Topic $errorTopic does not exist, createTopicIfNotExists == false")
           throw new IllegalArgumentException(s"Errors topic: $errorTopic does not exist, 'createTopicIfNotExists' is 'false', please create the topic")